dtype: assert the float64 the barrier cap assumes, and do the autocast sweep
Two halves of review 2.14.2.

(1) THE ASSUMPTION NOBODY CHECKED.  potentials.py caps every soft bound's
log-sigmoid argument at _MAX_ARG = 700 because exp(700) ~ 1e304 is finite in
float64.  Nothing asserted the float64.  Under floatX=float32 that same graph
overflows exp at ~88 and underflows sigmoid to exactly 0 far inside the region
where float64 still has a usable gradient, so every soft barrier would
evaluate to -inf on its forbidden side -- a wall with no restoring force, and
no message anywhere.  exozippy/__init__.py now refuses to import, right beside
the linker fix-up, naming potentials.py and the two places floatX can be set.
A RuntimeError rather than an `assert`, so `python -O` cannot strip it.

(2) THE SWEEP THAT WAS NEVER RUN.  pytensor autocasts a bare Python float to
the SMALLEST dtype that represents it, so pt.as_tensor_variable(5778.0) is
float32 -- the trap docs/testing.md documents precisely because it keeps being
rediscovered.  Three live sites:

  parameter.py, the truncated-normal mass correction: the Gaussian prior
    CENTRE was `pt.as_tensor_variable(float(prior_mus[i]))`, i.e. a float32
    constant, so a logp term carried a mu rounded to ~1e-7 relative.  The only
    one of the three where a VALUE was wrong.  -> np.float64.
  parameter.py, the fixed/derived raw elements: `pt.constant(0.0)`.  0.0 is
    exact in float32 and the stack upcasts while any element is sampled, so
    this was invisible -- but a vector whose elements are ALL fixed or derived
    came out float32 end to end.  -> explicit dtype.
  transit.py, the undiluted-band placeholder: `pt.constant(1.0)`, same shape.
    -> explicit dtype.

Everything else reaching as_tensor_variable/constant in the tree passes a
numpy array or a numpy scalar, which keeps float64.

Tests (tests/test_dtype_consistency.py): the float32 import really refuses (in
a subprocess -- floatX is read at import); exp(_MAX_ARG) is finite in float64
and not in float32, stated as a fact rather than as prose; the autocast
behavior itself is pinned so this cannot be rediscovered a fourth time
silently; and a source scan forbids a bare float literal reaching either
constructor.

NOT FIXED, DELIBERATELY: components/mulensing/mulensinstrument.py:1618's
`pt.constant(0.0)` is the same defect, but that file is under an active
rewrite for item 8.6.7 and was out of scope here.  It is a dtype-only issue
(0.0 is exact, and it upcasts in the stack it feeds).  The source scan carries
it as a named, explained allowlist entry rather than being narrowed, so it
stays visible and the next person to touch that file removes the entry.

Item 2.14.2.

+++ src/exozippy/components/parameter.py
@@ -1821,9 +1821,13 @@ class Parameter:
         # 4. BUILD RAW VARIABLES
         raw_elements = [None] * n_elements
 
-        # Fixed / derived: constant 0 in raw space
+        # Fixed / derived: constant 0 in raw space.  dtype is explicit
+        # because a bare 0.0 autocasts to float32 and these are stacked with
+        # the sampled elements -- harmless while at least one element is
+        # sampled (the stack upcasts), but a vector whose elements are ALL
+        # fixed or derived would come out float32 end to end (review 2.14.2).
         for i in np.where(is_fixed | is_derived)[0]:
-            raw_elements[i] = pt.constant(0.0)
+            raw_elements[i] = pt.constant(0.0, dtype="float64")
 
         if np.any(is_sampled):
             idx = np.where(is_sampled)[0]
@@ -2220,7 +2224,12 @@ class Parameter:
             if i in mu_links:
                 mu_i = mu_links[i]["fn"](val_flat)
             elif gaussian_prior_mask[i]:
-                mu_i = pt.as_tensor_variable(float(prior_mus[i]))
+                # np.float64, NOT a bare Python float: pytensor autocasts
+                # a Python float to the SMALLEST dtype that represents it,
+                # so `float(...)` here produced a float32 constant and this
+                # truncation-mass logp term carried a mu rounded to ~1e-7
+                # relative (docs/testing.md's autocast trap; review 2.14.2).
+                mu_i = pt.as_tensor_variable(np.float64(prior_mus[i]))
             else:
                 continue  # no Gaussian on this element: U(lo, up) already
             sig_i = float(sigmas[i])
